+++ src/utils/dateUtils.ts
@@ -108,3 +108,84 @@ export function formatDate(currentDate: Date, day?: number) {
     fillZero(day ?? currentDate.getDate()),
   ].join('-');
 }
+
+// 주어진 날짜에 interval일을 더한 날짜를 반환
+export function addDays(date: Date, interval: number): Date {
+  const newDate = new Date(date);
+  newDate.setDate(date.getDate() + interval);
+  return newDate;
+}
+
+// 주어진 날짜에 interval주를 더한 날짜를 반환
+export function addWeeks(date: Date, interval: number): Date {
+  return addDays(date, interval * 7);
+}
+
+// 주어진 날짜에 interval개월을 더한 날짜를 반환
+export function addMonths(date: Date, interval: number): Date {
+  const newDate = new Date(date);
+  const targetMonth = newDate.getMonth() + interval;
+  const targetYear = newDate.getFullYear() + Math.floor(targetMonth / 12);
+  const adjustedMonth = targetMonth % 12;
+
+  newDate.setFullYear(targetYear);
+  newDate.setMonth(adjustedMonth);
+
+  // 원래 날짜가 31일이었는데 30일까지인 달로 이동한 경우 등
+  // 유효하지 않은 날짜가 있으면 해당 달을 건너뛰고 다음 달로 이동
+  if (newDate.getDate() !== date.getDate()) {
+    // 해당 달을 건너뛰고 다음 달의 같은 날짜로 설정
+    newDate.setMonth(adjustedMonth + 1);
+    newDate.setDate(date.getDate());
+
+    // 다음 달도 유효하지 않으면 계속 다음 달로 이동
+    while (newDate.getDate() !== date.getDate()) {
+      newDate.setMonth(newDate.getMonth() + 1);
+      newDate.setDate(date.getDate());
+    }
+  }
+
+  return newDate;
+}
+
+// 윤년인지 확인하는 함수
+function isLeapYear(year: number): boolean {
+  return (year % 4 === 0 && year % 100 !== 0) || year % 400 === 0;
+}
+
+// 주어진 날짜에 interval년을 더한 날짜를 반환
+export function addYears(date: Date, interval: number): Date {
+  // 2월 29일인 경우 특별 처리
+  if (date.getMonth() === 1 && date.getDate() === 29) {
+    let targetYear = date.getFullYear() + interval;
+
+    // 윤년이 아닌 해는 건너뛰고 다음 윤년으로 이동
+    while (!isLeapYear(targetYear)) {
+      targetYear++;
+    }
+
+    // 새로운 Date 객체를 생성하여 2월 29일로 설정
+    return new Date(targetYear, 1, 29);
+  }
+
+  // 일반적인 경우
+  const newDate = new Date(date);
+  newDate.setFullYear(date.getFullYear() + interval);
+  return newDate;
+}
+
+// 반복 타입에 따라 다음 날짜를 계산
+export function getNextDate(currentDate: Date, repeatType: string, interval: number): Date {
+  switch (repeatType) {
+    case 'daily':
+      return addDays(currentDate, interval);
+    case 'weekly':
+      return addWeeks(currentDate, interval);
+    case 'monthly':
+      return addMonths(currentDate, interval);
+    case 'yearly':
+      return addYears(currentDate, interval);
+    default:
+      return currentDate;
+  }
+}
